Change type from double to decimal for Attack and Health Hero attributes

+++ rpgsaga/src/RpgSaga.Core/Abstractions/HeroFactory.cs
@@ -2,4 +2,4 @@
 
 namespace RpgSaga.Core.Abstractions;
 
-public delegate Hero HeroFactory(string firstName, string lastName, double health, double attack);
+public delegate Hero HeroFactory(string firstName, string lastName, decimal health, decimal attack);
