+++ rpgsaga/src/RpgSaga.DLCs/Rogue/Heroes/Mage.cs
@@ -4,12 +4,12 @@ namespace RpgSaga.DLCs.Rogue.Heroes;
 
 public sealed class Mage : Hero
 {
-    public Mage(string firstName, string lastName, double health, double attack) 
+    public Mage(string firstName, string lastName, decimal health, decimal attack) 
         : base(firstName, lastName, health, attack)
     {
     }
 
-    public static Mage Create(string firstName, string lastName, double health, double attack)
+    public static Mage Create(string firstName, string lastName, decimal health, decimal attack)
     {
         return new Mage(firstName, lastName, health, attack);
     }
